requirements.txt
aiogram==3.3.0
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic>=2.4.1,<2.6.0
email_validator>=2.0.0
google-api-python-client==2.117.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.7
jinja2==3.1.2
